requirements.txt
lxml==5.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-xmltv ~= 1.4.3
fake-useragent==1.4.0
python-dateutil ~= 2.8.2
PyYAML ~= 6.0
py-xmltv ~= 1.0.8
roman==4.1
tqdm==4.66.1
flask==3.0.0
schedule==1.2.1
loguru==0.7.2
pendulum==3.0.0

